backend/requirements.txt
asgiref==3.7.2
black==23.11.0
Django==5.0
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Pillow==10.1.0
referencing==0.31.0
django-cors-headers==4.3.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
Faker==20.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
aiohttp==3.9.1
channels==4.0.0